PinnedConcurrentCol migration
`ConcurrentVec` now is only a wrapper around a `ConcurrentBag` with safe get and iter implementations.
* `ConcurrentBag` implements `Deref` and `DerefMut` for `PinnedConcurrentCol `. And `ConcurrentVec` for `ConcurrentBag `.
* Documentation is revised.
* Benchmarks are repeated, performance has not changed (as expected).
