+++ src/lib.rs
@@ -3,15 +3,14 @@
 //! [![orx-concurrent-vec crate](https://img.shields.io/crates/v/orx-concurrent-vec.svg)](https://crates.io/crates/orx-concurrent-vec)
 //! [![orx-concurrent-vec documentation](https://docs.rs/orx-concurrent-vec/badge.svg)](https://docs.rs/orx-concurrent-vec)
 //!
-//! An efficient, convenient and lightweight grow-only and read-and-write concurrent collection.
+//! An efficient, convenient and lightweight grow-only read & write concurrent data structure allowing high performance concurrent collection.
 //!
-//! * **convenient**: `ConcurrentVec` can safely be shared among threads simply as a shared reference. Further, it is a wrapper around any [`PinnedVec`](https://crates.io/crates/orx-pinned-vec) implementation adding concurrent safety guarantees. Underlying pinned vector of `Option<T>` and concurrent vec of `T` can be converted to each other back and forth without any cost (see <a href="#section-construction-and-conversions">construction and conversions</a>).
-//! * **lightweight**: This crate takes a simplistic approach built on pinned vector guarantees which leads to concurrent programs with few dependencies and small binaries (see <a href="#section-approach-and-safety">approach and safety</a> for details).
-//! * **efficient**: `ConcurrentBag` is a grow only collection with immutable elements making use of atomic primitives. This leads to efficient growth and lock-free reading. You may see the details in <a href="#section-benchmarks">benchmarks</a> and further <a href="#section-performance-notes">performance notes</a>.
+//! * **convenient**: `ConcurrentVec` can safely be shared among threads simply as a shared reference. It is a [`PinnedConcurrentCol`](https://crates.io/crates/orx-pinned-concurrent-col) with a special concurrent state implementation. Underlying [`PinnedVec`](https://crates.io/crates/orx-pinned-vec) and concurrent bag can be converted back and forth to each other.
+//! * **efficient**: `ConcurrentVec` is a lock free structure making use of a few atomic primitives, this leads to high performance concurrent growth. You may see the details in <a href="#section-benchmarks">benchmarks</a> and further <a href="#section-performance-notes">performance notes</a>.
 //!
-//! Note that `ConcurrentVec` is a **read and write** collection. When we only need to collect results concurrently without reading, [`ConcurrentBag`](https://crates.io/crates/orx-concurrent-bag) is preferable due to its performance optimizations. Having almost identical api, switching between `ConcurrentVec` and `ConcurrentBag` is straightforward.
+//! Note that `ConcurrentVec` is a read & write collection with the cost to store values wrapped with an optional and initializing memory on allocation. See [`ConcurrentBag`](https://crates.io/crates/orx-concurrent-bag) for a write only and a more performant variant. Having almost identical api, switching between `ConcurrentVec` and `ConcurrentBag` is straightforward.
 //!
-//! # A. Examples
+//! ## Examples
 //!
 //! The main feature of `ConcurrentVec` compared to concurrent bag is to enable safe reading while providing efficient growth. It is convenient to share the concurrent vector among threads. `std::sync::Arc` can be used; however, it is not required as demonstrated below.
 //!
@@ -100,89 +99,7 @@
 //! assert_eq!(metrics.len(), 10);
 //! ```
 //!
-//! <div id="section-approach-and-safety"></div>
-//!
-//! # B. Approach and Safety
-//!
-//! `ConcurrentVec` is a wrapper around the `ConcurrentBag` with a focus to enable safe concurrent reading while concurrent bag focuses on efficient growth. Details of the approach and growth related safety guarantees can be found [here](https://docs.rs/orx-concurrent-bag/latest/orx_concurrent_bag/#section-approach-and-safety).
-//!
-//! In order to add safe concurrent reading feature to the concurrent bag, the following simple approach is taken:
-//! * When `ConcurrentBag` allocates to grow, it does not initialize the memory of positions which are less than the length (as common in dynamic size vectors). Since reading from a concurrent bag only happens after the consuming `into_inner` method, this is perfectly fine. Read methods such as `get(index)` and `iter()` are available; however, unsafe, due to the possibility of the following sequence of events:
-//!   * `bag.push('a')` is called from thread#1.
-//!   * `bag` atomically increases the `len` to 1.
-//!   * thread#2 calls `bag.get(0)` which is now in bounds.
-//!   * thread#2 receives uninitialized value which is an undefined behavior (UB).
-//!   * thread#1 completes writing `'a'` to the 0-th position (one moment too late).
-//! * `ConcurrentVec` prevents this as follows:
-//!   * Instead of `T`, elements of a concurrent vector are stored as `Option<T>`.
-//!   * When `ConcurrentVec` allocates, it initializes all new positions as `None`.
-//! * Read methods such as `get` or `iter` skips in-bounds but `None` elements. These are the elements which are reserved by some threads; however, the process of writing them is not completed yet.
-//! * Further, elements already pushed to the concurrent vector are immutable. In other words, if the value at a position is of `Some` variant, it will never change.
-//! * Therefore, it is possible to read already pushed elements without a race condition.
-//!
-//! <div id="section-benchmarks"></div>
-//!
-//! # C. Benchmarks
-//!
-//! ## Performance with `push`
-//!
-//! *You may find the details of the benchmarks at [benches/collect_with_push.rs](https://github.com/orxfun/orx-concurrent-vec/blob/main/benches/collect_with_push.rs).*
-//!
-//! In the experiment, `rayon`s parallel iterator and `ConcurrentVec`s `push` method are used to collect results from multiple threads.
-//!
-//! ```rust ignore
-//! // reserve and push one position at a time
-//! for j in 0..num_items_per_thread {
-//!     bag_ref.push(i * 1000 + j);
-//! }
-//! ```
-//!
-//! * We observe that rayon is significantly faster when the output is very small (`i32` in this experiment).
-//! * As the output gets larger and copies become costlier (`[i32; 32]` here), `ConcurrentVec::push` starts to perform equivalent to or faster than rayon.
-//! * Among the `ConcurrentVec` variants, `Linear` and `Fixed` variants perform faster than the `Doubling` variant:
-//!   * Here we observe the cost of memory initialization immediately on allocation. Since `Doubling` variant allocates more, initialization has a greater impact.
-//!   * `ConcurrentBag` does not perform this operation and it leads to a very high performance concurrent collection. Further, the impact of the underlying pinned vector type is insignificant. Therefore, it is a better choice when we only write results concurrently.
-//!   * The main goal of `ConcurrentVec`, on the other hand, is to enable safe reading while the vector concurrently grows, and it must be preferred in these situations over making unsafe calls.
-//!   * Having almost identical api, switching between bag and vec is straightforward.
-//!
-//! The issue leading to poor performance in the *small data & little work* situation can be avoided by using `extend` method in such cases. You may see its impact in the succeeding subsections and related reasons in the <a href="#section-performance-notes">performance notes</a>.
-//!
-//! <img src="https://raw.githubusercontent.com/orxfun/orx-concurrent-vec/main/docs/img/bench_collect_with_push.PNG" alt="https://raw.githubusercontent.com/orxfun/orx-concurrent-vec/main/docs/img/bench_collect_with_push.PNG" />
-//!
-//! ## Performance of `extend`
-//!
-//! *You may find the details of the benchmarks at [benches/collect_with_extend.rs](https://github.com/orxfun/orx-concurrent-vec/blob/main/benches/collect_with_extend.rs).*
-//!
-//! The only difference in this follow up experiment is that we use `extend` rather than `push` with `ConcurrentVec`. The expectation is that this approach will solve the performance degradation due to false sharing in the *small data & little work* situation.
-//!
-//! ```rust ignore
-//! // reserve num_items_per_thread positions at a time
-//! // and then push as the iterator yields
-//! let iter = (0..num_items_per_thread).map(|j| i * 100000 + j);
-//! bag_ref.extend(iter);
-//! ```
-//!
-//! <img src="https://raw.githubusercontent.com/orxfun/orx-concurrent-vec/main/docs/img/bench_collect_with_extend.PNG" alt="https://raw.githubusercontent.com/orxfun/orx-concurrent-vec/main/docs/img/bench_collect_with_extend.PNG" />
-//!
-//! Note that we do not need to have perfect information on the number of items to be pushed per thread to get the benefits of `extend`, we can simply `step_by`. Extending by `batch_size` elements will already prevent the dramatic performance degradation provided that `batch_size` elements exceed a cache line.
-//!
-//! ```rust ignore
-//! // reserve batch_size positions at a time
-//! // and then push as the iterator yields
-//! for j in (0..num_items_per_thread).step_by(batch_size) {
-//!     let iter = (j..(j + batch_size)).map(|j| i * 100000 + j);
-//!     bag_ref.extend(iter);
-//! }
-//! ```
-//!
-//! <div id="section-performance-notes"></div>
-//!
-//! # D. Performance Notes
-//!
-//! `ConcurrentVec` is an efficient read-and-write collection. However, it is important to avoid false sharing risk which might lead to a significant performance degradation. Details can be read [here](https://docs.rs/orx-concurrent-bag/latest/orx_concurrent_bag/#section-performance-notes).
-//! <div id="section-construction-and-conversions"></div>
-//!
-//! # E. `From` | `into_inner`
+//! ### Construction
 //!
 //! `ConcurrentVec` can be constructed by wrapping any pinned vector; i.e., `ConcurrentVec<T>` implements `From<P: PinnedVec<Option<T>>>`.
 //! Likewise, a concurrent vector can be unwrapped to get the underlying pinned vector with `into_inner` method.
@@ -222,7 +139,79 @@
 //! let con_vec: ConcurrentVec<_> = split_vec.into();
 //! ```
 //!
-//! # License
+//! ## Concurrent State and Properties
+//!
+//! The concurrent state is modeled simply by an atomic length. Combination of this state and `PinnedConcurrentCol` leads to the following properties:
+//! * Writing to a position of the collection does not block other writes, multiple writes can happen concurrently.
+//! * Each position is written exactly once.
+//! * ⟹ no write & write race condition exists.
+//! * Only one growth can happen at a given time.
+//! * Underlying pinned vector is always valid and can be taken out any time by `into_inner(self)`.
+//! * Reading a position while its being written will yield `None` and will be omitted.
+//!
+//! <div id="section-benchmarks"></div>
+//!
+//! ## Benchmarks
+//!
+//! ### Performance with `push`
+//!
+//! *You may find the details of the benchmarks at [benches/collect_with_push.rs](https://github.com/orxfun/orx-concurrent-vec/blob/main/benches/collect_with_push.rs).*
+//!
+//! In the experiment, `rayon`s parallel iterator and `ConcurrentVec`s `push` method are used to collect results from multiple threads.
+//!
+//! ```rust ignore
+//! // reserve and push one position at a time
+//! for j in 0..num_items_per_thread {
+//!     bag_ref.push(i * 1000 + j);
+//! }
+//! ```
+//!
+//! * We observe that rayon is significantly faster when the output is very small (`i32` in this experiment).
+//! * As the output gets larger and copies become costlier (`[i32; 32]` here), `ConcurrentVec::push` starts to perform equivalent to or faster than rayon.
+//! * Among the `ConcurrentVec` variants, `Linear` and `Fixed` variants perform faster than the `Doubling` variant:
+//!   * Here we observe the cost of memory initialization immediately on allocation. Since `Doubling` variant allocates more, initialization has a greater impact.
+//!   * `ConcurrentBag` does not perform this operation and it leads to a very high performance concurrent collection. Further, the impact of the underlying pinned vector type is insignificant. Therefore, it is a better choice when we only write results concurrently.
+//!   * The main goal of `ConcurrentVec`, on the other hand, is to enable safe reading while the vector concurrently grows, and it must be preferred in these situations over making unsafe calls.
+//!   * Having almost identical api, switching between bag and vec is straightforward.
+//!
+//! The issue leading to poor performance in the *small data & little work* situation can be avoided by using `extend` method in such cases. You may see its impact in the succeeding subsections and related reasons in the <a href="#section-performance-notes">performance notes</a>.
+//!
+//! <img src="https://raw.githubusercontent.com/orxfun/orx-concurrent-vec/main/docs/img/bench_collect_with_push.PNG" alt="https://raw.githubusercontent.com/orxfun/orx-concurrent-vec/main/docs/img/bench_collect_with_push.PNG" />
+//!
+//! ### Performance of `extend`
+//!
+//! *You may find the details of the benchmarks at [benches/collect_with_extend.rs](https://github.com/orxfun/orx-concurrent-vec/blob/main/benches/collect_with_extend.rs).*
+//!
+//! The only difference in this follow up experiment is that we use `extend` rather than `push` with `ConcurrentVec`. The expectation is that this approach will solve the performance degradation due to false sharing in the *small data & little work* situation.
+//!
+//! ```rust ignore
+//! // reserve num_items_per_thread positions at a time
+//! // and then push as the iterator yields
+//! let iter = (0..num_items_per_thread).map(|j| i * 100000 + j);
+//! bag_ref.extend(iter);
+//! ```
+//!
+//! <img src="https://raw.githubusercontent.com/orxfun/orx-concurrent-vec/main/docs/img/bench_collect_with_extend.PNG" alt="https://raw.githubusercontent.com/orxfun/orx-concurrent-vec/main/docs/img/bench_collect_with_extend.PNG" />
+//!
+//! Note that we do not need to have perfect information on the number of items to be pushed per thread to get the benefits of `extend`, we can simply `step_by`. Extending by `batch_size` elements will already prevent the dramatic performance degradation provided that `batch_size` elements exceed a cache line.
+//!
+//! ```rust ignore
+//! // reserve batch_size positions at a time
+//! // and then push as the iterator yields
+//! for j in (0..num_items_per_thread).step_by(batch_size) {
+//!     let iter = (j..(j + batch_size)).map(|j| i * 100000 + j);
+//!     bag_ref.extend(iter);
+//! }
+//! ```
+//!
+//! <div id="section-performance-notes"></div>
+//!
+//! ### Performance Notes
+//!
+//! `ConcurrentVec` is an efficient read-and-write collection. However, it is important to avoid false sharing risk which might lead to a significant performance degradation. Details can be read [here](https://docs.rs/orx-concurrent-bag/latest/orx_concurrent_bag/#section-performance-notes).
+//! <div id="section-construction-and-conversions"></div>
+//!
+//! ## License
 //!
 //! This library is licensed under MIT license. See LICENSE for details.
 
@@ -238,6 +227,7 @@
     clippy::todo
 )]
 
+mod new;
 mod vec;
 
 pub use orx_fixed_vec::FixedVec;
